tweak(web): TAM-2394: move AI form builder into feature folder
