+++ packages/web/app/views/administration/programs/surveys/aiFormBuilder/index.js
@@ -1,0 +1,1 @@
+export { AiFormBuilderView } from './AiFormBuilderView';
