+++ packages/web/app/views/administration/programs/surveys/aiFormBuilder/mockGenerateForm.js
