+++ packages/web/app/views/administration/programs/surveys/index.js
@@ -1,3 +1,3 @@
-export { AiFormBuilderView } from './AiFormBuilderView';
+export { AiFormBuilderView } from './aiFormBuilder';
 export { ManageProgramsAdminView } from './ManageProgramsAdminView';
 export { ProgramsAdminView, ProgramsExportTab, ProgramsImportTab } from './ProgramsAdminView';
